sources/requirements.txt
Django==5.1.4
psycopg2-binary==2.9.6
django-cors-headers==4.6
python-dotenv>=0.17
djangorestframework==3.15.2
django-allauth>=0.44
requests>=2.25
PyJWT>=2.0
cryptography>=3.4
paramiko>=2.7
Pillow>=11.1
daphne>=4.1.2
channels>=4.2.0
redis>=5.2.1
sendgrid>=6.11
djangorestframework-simplejwt==5.4